requirements.txt
#Django==1.8.2
django-bootstrap3==5.1.1
django-bootstrap3-datetimepicker
pinax-theme-bootstrap==5.7.6
ipdb==0.8.1
ipython==3.1.0
wheel==0.24.0
django-bootstrap3-datetimepicker==2.2.3

